run formatters, bandit and restructure
