--- python/dremio-flight-endpoint/middleware/__init__.py
@@ -1,1 +1,0 @@
-
