+++ python/dremio/middleware/auth_middleware.py
